Add files via upload
The data preprocessing and EDA as done. We trained decision tree, random forest, and XGBoost algorithms and choose the algorithm with the best performance. The accuracy score is 95.39% for the model_dt and the classification report also has above 90% for almost all except harmonics. In confusion metrics 170 records of flicker were correctly predicted as flicker, out of 170 records 149 was correctly predicted as harmonics, 170 was correctly predicted as an interruption, 170 was correctly predicted as normal, from 170 records 157 was correctly predicted as sag and from 170 records 157 was correctly predicted as swell. The accuracy score, classification report, and confusion metrics for XG Boost are the same as the decision tree. The accuracy score for the random forest is 97.75%. highest compared all the other models, which is clearly visible in the bar chart so Random Forest was finalized.